skyspy_django/requirements.txt
# Shared SkysPy packages
-e ../skyspy_common

# CFFI for better libacars performance (optional)
cffi>=1.15.0

# Django and REST Framework
Django>=5.0
djangorestframework>=3.14
drf-spectacular>=0.27
django-filter>=23.5
django-cors-headers>=4.3

# Authentication
djangorestframework-simplejwt>=5.3
mozilla-django-oidc>=4.0
PyJWT>=2.8

# Real-time (Django Channels)
channels>=4.0
channels-redis>=4.1
daphne>=4.0
django-eventstream>=4.5

# Background Tasks (Celery)
celery[redis,gevent]>=5.3
django-celery-beat>=2.5
gevent>=24.2

# Database
psycopg[binary]>=3.1
asyncpg>=0.29.0

# HTTP Client
httpx>=0.26.0

# Notifications
apprise>=1.6

# Cloud Storage
boto3>=1.34

# Monitoring
prometheus-client>=0.19
sentry-sdk>=1.39

# Utilities
python-dateutil>=2.8
pytz>=2024.1

# Audio processing
mutagen>=1.47  # Audio metadata/duration extraction

# Production servers (optional alternatives)
gunicorn>=21.0
uvicorn[standard]>=0.27

# Static file serving
whitenoise>=6.6
